==> src/notebooks/1_api_kaggle.ipynb <==
# Importando biblioteca kagglehub

import kagglehub
import os, shutil

# Baixando o arquivo do Kaggle

path = kagglehub.dataset_download("parisrohan/credit-score-classification")
print("Path to dataset files:", path)

# Movendo o arquivo do Kaggle para pasta do projeto

for i in os.listdir(path):
    if i.endswith('.csv'):
        shutil.move(os.path.join(path, i), os.path.join(r'C:\Users\GabGalani\Documents\Data Projects\Analise_credito\dataset\bonze', i))
        print(f'Arquivo movido - {i}')
