==> src/notebooks/2_analise_exploratoria.ipynb <==
# Importando bibliotecas

import pandas as pd
import numpy as np
import os, re
from pathlib import Path

pd.set_option('display.max_columns', 90)
pd.set_option('display.max_rows', 100)

# Variáveis

dataset_folder = r'C:\Users\GabGalani\Documents\Data Projects\Analise_credito\dataset\bronze'
file_test = 'test.csv'

# Datasets

## Dataset test.csv

dataset = pd.read_csv(os.path.join(dataset_folder, file_test), sep=',')
dataset.head()

dataset.info()

## Verificando e tratando nulos

dataset.isnull().sum()

- Coluna Name nula, deverá ser preenchida, pois há casos como o acima, onde o ID já possui o nome; (Essa é irrelevante)
- Monthly_Inhand_Salary é importante para avaliar a capacidade de pagamento, então trataei os nulos com a mediana da coluna       Annual_Income;
- Num_of_Delayed_Payment é importante pois avalia o comportamento do pagamento, então assumiremos como 0;
- Num_Credit_Inquiries é importante pois mostra quantas consultas de crédito foram feitas recentemente, vamos considerar como 0;
- Credit_History_Age é importante pois indica o histórico de crédito, vamos considerar fazer um dataset para verificar se já possui essa informação em outra linha e agregar calculando o tempo
- Amount_invested_monthly Indica quanto o cliente investe mensalmente, vamos conseidarar 0 para os nulos devido à cada linha ser um mês o nulo indica que não houve depósito

## Tratando a coluna Age

dataset.head(2)

dataset['Age'] = dataset['Age'].str.replace('_', '', regex=False).astype(float)
dataset['Age'] = dataset['Age'].astype(int)


## Coluna Name

dataset_name = dataset[['Customer_ID', 'Name']]
dataset_name = dataset_name[(~dataset_name['Name'].isnull()) & (~dataset_name['Customer_ID'].isnull())]
dataset_name = dataset_name.drop_duplicates()

dataset = dataset.merge(dataset_name, on='Customer_ID', how='left', suffixes=('', '_new'))
dataset['Name'] = dataset['Name'].fillna(dataset['Name_new'])
dataset['Name'] = dataset['Name'].fillna(dataset['Name_new'])

## Coluna Credit_History_Age

def convert_to_months(age_str):
    if pd.isnull(age_str):
        return None
    parts = age_str.split(' ')
    years = int(parts[0])
    months = int(parts[3])
    return years * 12 + months

def convert_to_years_months(months):
    if pd.isnull(months):
        return "0 Years and 0 Months"  # Retornar um valor padrão para NaN
    years = months // 12
    rem_months = months % 12
    return f"{int(years)} Years and {int(rem_months)} Months"


dataset_credit_history = dataset[['Customer_ID', 'Month', 'Age', 'Credit_History_Age']]

dataset_credit_history['Month_number'] = pd.to_datetime(dataset_credit_history['Month'], format='%B')
dataset_credit_history = dataset_credit_history.sort_values(by=['Customer_ID', 'Month'])

# Converte o 'Credit_History_Age' para 'Credit_History_Months'
dataset_credit_history['Credit_History_Months'] = dataset_credit_history['Credit_History_Age'].apply(convert_to_months)

# Preencher valores nulos
dataset_credit_history['Credit_History_Months'] = dataset_credit_history['Credit_History_Months'].fillna(dataset_credit_history['Credit_History_Months'].shift(1) - 1)

# Preencher valores nulos restantes usando bfill
dataset_credit_history['Credit_History_Months'] = dataset_credit_history.groupby('Customer_ID')['Credit_History_Months'].bfill()

# Convertendo novamente para o formato 'Years and Months'
dataset_credit_history['Credit_History_Age'] = dataset_credit_history['Credit_History_Months'].apply(convert_to_years_months)

# Exibindo os resultados
dataset_credit_history.head(2)


Ajustando a coluna tratada

dataset = dataset.merge(dataset_credit_history, on=['Customer_ID', 'Month', 'Age'], how='left', suffixes=('', '_new'))
dataset['Credit_History_Age'] = dataset['Credit_History_Age'].fillna(dataset['Credit_History_Age_new'])
dataset['Credit_History_Age'] = dataset['Credit_History_Age'].fillna(dataset['Credit_History_Age_new'])

def extract_year(age_str):
    # Verificar se age_str é uma string antes de aplicar split
    if isinstance(age_str, str):
        parts = age_str.split(' ')
        years = int(parts[0])
        return years+1
    return 0  # Retornar None caso não seja uma string


dataset['Age_new'] = dataset['Credit_History_Age'].apply(extract_year).astype(int)

dataset['Age'] = np.where(
    (dataset['Age'] >= 100) | (dataset['Age'] < 0),
    dataset['Age_new'],
    dataset['Age']
)

## Monthly_Inhand_Salary 

Importante para avaliar capacidade de pagamento. <br>
Preencher com a mediana ou calcular com base na Annual_Income

dataset['Annual_Income'] = dataset['Annual_Income'].str.replace('_', '', regex=False).astype(float)
dataset['Annual_Income'] = dataset['Annual_Income'].astype(int)
dataset['Annual_Income'].astype(float)
dataset['Monthly_Inhand_Salary'] = dataset['Monthly_Inhand_Salary'].fillna(dataset['Annual_Income'] / 12)

## Num_of_Delayed_Payment

Avalia comportamento de pagamento. <br>
Preencher com 0 (assumir que não houve atraso) ou com a mediana

dataset['Num_of_Delayed_Payment'] = dataset['Num_of_Delayed_Payment'].fillna(0)

## Num_Credit_Inquiries

Mostra quantas consultas ao crédito foram feitas (uso frequente pode ser sinal de risco). <br>
Preencher com a mediana ou 0:

dataset['Num_Credit_Inquiries'] = dataset['Num_Credit_Inquiries'].fillna(0)


## Amount_invested_monthly

Indica quanto o cliente investe por mês, relevante para estabilidade financeira. <br>
Preencher com 0 para o mês nulo


dataset['Amount_invested_monthly'] = dataset['Amount_invested_monthly'].fillna(0)

## Monthly_Balance

Saldo mensal restante após despesas e pagamentos. <br>
Preencher com mediana

dataset['Monthly_Balance'] = pd.to_numeric(dataset['Monthly_Balance'], errors='coerce')
dataset['Monthly_Balance'] = dataset['Monthly_Balance'].fillna(dataset['Monthly_Balance'].mean())

## Type_of_Loan 

Pode indicar o tipo de empréstimo obtido (ex.: pessoal, carro, etc.). <br>
|Preencher com "Unknown"

dataset['Type_of_Loan'] = dataset['Type_of_Loan'].fillna('')

## Colunas numericas e float

int_columns = ['Age', 'Num_Bank_Accounts', 'Num_Credit_Card', 'Num_of_Loan', 'Delay_from_due_date', 'Num_of_Delayed_Payment', 'Num_Credit_Inquiries']
float_columns = ['Annual_Income', 'Monthly_Inhand_Salary', 'Interest_Rate', 'Changed_Credit_Limit', 'Outstanding_Debt', 'Credit_Utilization_Ratio', 'Total_EMI_per_month', 'Amount_invested_monthly', 'Monthly_Balance']

Transformando todo dataset em string

for i in dataset.columns: 
    dataset[i] = dataset[i].astype(str)

Colunas int

for i in int_columns:
    dataset[i] = dataset[i].str.replace('_', '', regex=False).astype(float)
    dataset[i] = dataset[i].astype(int)

Colunas float

dataset['Changed_Credit_Limit'] = dataset['Changed_Credit_Limit'].replace('_', '0')
float_columns = ['Annual_Income', 
                 'Monthly_Inhand_Salary',
                 'Interest_Rate', 
                 'Changed_Credit_Limit', 
                 'Outstanding_Debt', 
                 'Credit_Utilization_Ratio', 
                 'Total_EMI_per_month', 
                 'Amount_invested_monthly', 
                 'Monthly_Balance'
            ]

for i in float_columns:
    dataset[i] = dataset[i].str.replace('_', '', regex=False).astype(float)
    dataset[i] = dataset[i].astype(float)

## Verificando as colunas tratadas

dataset.isnull().sum()

## Verificando valor _______ na coluna Occupation

dataset['Occupation'].unique()

dataset[dataset['Occupation'] == '_______'].head(2)

dataset[dataset['Customer_ID'] == 'CUS_0x21b1'].head(2)

dataset_occupation = dataset[['Customer_ID', 'Occupation']]
dataset_occupation = dataset_occupation[dataset_occupation['Occupation'] != "_______"]
dataset_occupation = dataset_occupation.drop_duplicates()

dataset = dataset.merge(dataset_occupation, on='Customer_ID', how="left", suffixes=('', '_new'))
dataset['Occupation'] = np.where(
    dataset['Occupation'] == '_______',
    dataset['Occupation_new'],
    dataset['Occupation']
)

dataset['Occupation'].unique()

## Tratando valor !@9#%8 na coluna Payment_Behaviour

dataset['Payment_Behaviour'].unique()

dataset[dataset['Payment_Behaviour'] == '!@9#%8'].head(2)

dataset[dataset['Customer_ID'] == 'CUS_0x21b1'].head()

def classify_payment_behaviour(row):
    total_spent = row['Total_EMI_per_month'] + row['Amount_invested_monthly'] + row['Outstanding_Debt']
    salary = row['Monthly_Inhand_Salary']
    
    if total_spent > 0.6 * salary:
        spending = 'High_spent'
    else:
        spending = 'Low_spent'
    
    if total_spent < 500:
        payment_value = 'Small_value_payments'
    elif total_spent < 2000:
        payment_value = 'Medium_value_payments'
    else:
        payment_value = 'Large_value_payments'

    return f'{spending}_{payment_value}'

dataset_teste = dataset
dataset_teste['Payment_Behaviour_new'] = dataset_teste.apply(classify_payment_behaviour, axis=1)
dataset['Payment_Behaviour'] = np.where(
    dataset['Payment_Behaviour'] == '!@9#%8',
    dataset['Payment_Behaviour_new'],
    dataset['Payment_Behaviour']
)

dataset['Payment_Behaviour'].unique()

## Tratando valores _ na coluna Credit_Mix

dataset['Credit_Mix'].unique()

dataset = dataset.sort_values(by=['Customer_ID', 'Month', 'Age'])
dataset['Credit_Mix_new'] = dataset['Credit_Mix'].replace('_', None).fillna(method='bfill')

dataset['Credit_Mix'] = np.where(
    dataset['Credit_Mix'] == '_',
    dataset['Credit_Mix_new'],
    dataset['Credit_Mix']
)
dataset['Credit_Mix'].unique()

## Tratando valores errados coluna Num_of_Loan

dataset['Num_of_Loan_new'] = dataset['Type_of_Loan'].apply(lambda x: len(x.split(',')))
dataset['Num_of_Loan'] = np.where(
    dataset['Type_of_Loan'] == 'Not Specified',
    1,
    dataset['Num_of_Loan']
)

dataset['Num_of_Loan'] = np.where(
    dataset['Type_of_Loan'] != dataset['Num_of_Loan_new'],
    dataset['Num_of_Loan_new'],
    dataset['Num_of_Loan']
)

dataset['Num_of_Loan'] = np.where(
    dataset['Type_of_Loan'] == '',
    0,
    dataset['Num_of_Loan']
)

## Excluindo as colunas _new

for i in dataset.columns:
    if "_new" in i or "Month_number" == i or 'Credit_History_Months' == i:
        dataset.drop(columns=i, inplace=True)

dataset.info()

# Criando a pasta silver e salvando o novo dataset

silver_folder = r'C:\Users\GabGalani\Documents\Data Projects\Analise_credito\dataset\silver'
Path(silver_folder).mkdir(parents=True, exist_ok=True)

dataset.to_csv(os.path.join(silver_folder, 'silver_test.csv'), sep=';', index=False)
