==> src/notebooks/4_preparando_graficos_dashbard.ipynb <==
# Importando dados

import pandas as pd
import os

# Lendo dataset

file_path = r'C:\Users\GabGalani\Documents\Data Projects\Analise_credito\dataset\silver\silver_test.csv'
gold_path= r'C:\Users\GabGalani\Documents\Data Projects\Analise_credito\dataset\gold'
dataset = pd.read_csv(file_path, sep=';')
dataset.head()

# Dados para gráficos

## Dashboard de credit_mix x Idade

dashboard_age = dataset[['Credit_Mix', 'Age']]

# Calculando a média de 'age' agrupada por 'Credit_Mix'
dashboard_age = dashboard_age.groupby('Credit_Mix')['Age'].mean().reset_index().round()

# Renomeando a coluna resultante para facilitar a leitura
dashboard_age.rename(columns={'Age': 'Media'}, inplace=True)

dashboard_age.head()

dashboard_age.to_csv(os.path.join(gold_path, 'dashboard_age.txt'), sep=';')

dataset.head()

## Dashboard Credit_mix x profissao

# Agrupando por 'Credit_Mix' e 'Occupation', contando as ocorrências de cada profissão em cada categoria de crédito
dashboard_profissao = dataset.groupby(['Credit_Mix', 'Occupation']).size().reset_index(name='Contagem')
dashboard_profissao.to_csv(os.path.join(gold_path, 'dashboard_profissao.txt'), sep=';')

dashboard_profissao.head()

dataset.head()

## Dashboard temporal

dashboard_tempo = dataset[['Credit_Mix', 'Month']]

dashboard_tempo = dashboard_tempo.groupby(['Month', 'Credit_Mix']).size().reset_index(name='Count')
dashboard_tempo.to_csv(os.path.join(gold_path, 'dashboard_tempo.txt'), sep=';')

dashboard_tempo.head()

dashboard_tempo['Month'].unique()

## Dados de comportamento de pagamento

dados_pagamento = dataset['Payment_Behaviour']
dados_pagamento.to_csv(os.path.join(gold_path, 'dados_pagamento.txt'), sep=';')
